CAN fix attempt
The comments with TODO:! represnt the differenet places where I have tried to put watchdog disabling. Seems like it doesn`t help or implementation is incrorrect. The watchdog are switched off and I can see that their LCK and ENDINIT bits are changing according to the code, but CCCR.INIT bit in CAN still can`t be changed. CCE bit is noramlly accesable